+++ packages/atmosphere/CHANGELOG.md
@@ -16,6 +16,7 @@
 - Relaxed the type of transmittance and irradiance textures from `DataTexture` to `Texture`.
 - Deprecated `sunIrradiance` and `skyIrradiance` options and props. Renamed them to `sunLight` and `skyLight`.
 - Deprecated `photometric` options and props. Now outputs luminance only.
+- Deprecated `IrradianceMask` and renamed it to `LightingMask`.
 - Migrated to core API changes.
 
 ## [0.12.0] - 2025-06-12